requirements.txt
edge-tts>=6.1.7
pywin32>=306; platform_system == "Windows"
pywebview>=4.0.0
